+++ rest/clients/jvm/src/main/kotlin/json/JsonUtils.kt
@@ -1,11 +1,11 @@
 package cc.getportal.sdk.json
 
+import cc.getportal.sdk.command.CommandWithId
+import cc.getportal.sdk.command.CommandWithIdSerializer
+import cc.getportal.sdk.command.Response
 import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
-import command.CommandWithId
-import command.CommandWithIdSerializer
-import command.Response
 
 object JsonUtils {
