+++ rest/clients/jvm/src/main/kotlin/model/Profile.kt
@@ -1,4 +1,4 @@
-package model
+package cc.getportal.sdk.model
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonProperty
